Refactor and define project transition and access classes
