+++ web/modules/projects/projects/projects.routing.yml
@@ -11,7 +11,7 @@ project.mediate:
     _title: 'Mediate Project'
     transition: 'mediate'
   requirements:
-    _custom_access: '\Drupal\projects\ProjectActionsAccess::accessProjectTransition'
+    _custom_access: '\Drupal\projects\ProjectTransitionAccess::accessProjectTransition'
     project: \d+
   options:
     _admin_route: TRUE
@@ -27,7 +27,7 @@ project.submit:
     _title: 'Submit Project'
     transition: 'submit'
   requirements:
-    _custom_access: '\Drupal\projects\ProjectActionsAccess::accessProjectTransition'
+    _custom_access: '\Drupal\projects\ProjectTransitionAccess::accessProjectTransition'
     project: \d+
   options:
     _admin_route: TRUE
